Rename map.py to map_instance.py
